Basic Functionality for the Page Completed
This is the first commit with my project being partially completed. The page I have made 'will' change the background color randomly BUT I have given it limited values in array.

So it won't be generating infinitely possible random backgrounds, instead it will only generate colors from the  data given in the array.